Make AMQP transport optional
The AMQP transport requires an external C library and extra installs.

Make it optional by building only when the tag "amqp" is set, also make the different
transports configure themselves by setting the entry in the transports map in the
init function.
